Python_Lab/CONFIG_NETCONF/requirement.txt
pyang
pyangbind
ncclient==0.6.7
json2xml
grpcio
grpcio-tools
paramiko
pyyaml

